requirements.txt
Pillow==5.1.0
tensorflow==1.7.0